Spli zustand store into feedback and business  store
